refactor: delete `AuthorizeRequestConfiguration` and `OidcUserInfoAugmentationConfiguration`, update refs to use CSL
