+++ security/security-core/src/main/java/io/camunda/security/oidc/CachingOidcClaimsProvider.java
@@ -10,8 +10,8 @@ package io.camunda.security.oidc;
 import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;
 import com.github.benmanes.caffeine.cache.Expiry;
+import io.camunda.security.api.model.config.oidc.OidcUserInfoAugmentationConfiguration;
 import io.camunda.security.configuration.OidcAuthenticationConfiguration;
-import io.camunda.security.configuration.OidcUserInfoAugmentationConfiguration;
 import io.camunda.zeebe.util.VisibleForTesting;
 import io.micrometer.core.instrument.Counter;
 import io.micrometer.core.instrument.MeterRegistry;
@@ -19,6 +19,7 @@ import io.micrometer.core.instrument.Timer;
 import java.net.URI;
 import java.time.Duration;
 import java.time.Instant;
+import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Objects;
